+++ src/commands/config/index.ts
@@ -1,0 +1,8 @@
+export * from "./config.ts";
+
+import { Command } from "../../deps.ts";
+import { config } from "./config.ts";
+
+export const configCmd = new Command()
+  .description("Configure global setting")
+  .action(config);
