+++ src/deps.ts
@@ -1,2 +1,3 @@
 export * from "https://deno.land/x/cliffy@v0.25.6/mod.ts";
 export * from "https://deno.land/std@0.170.0/fs/mod.ts";
+export { default as dir } from "https://deno.land/x/dir@1.5.1/mod.ts";
