--- src/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from "./general.helpers.ts";
-export * from "./validation.helpers.ts";
