+++ src/main.ts
@@ -1,12 +1,12 @@
-import configuration from "./config.json" assert { type: "json" };
+import config from "./config.json" assert { type: "json" };
 import { Command } from "./deps.ts";
 
-import { configCommand } from "./commands/index.ts";
+import { configCmd } from "./commands/index.ts";
 
 await new Command()
-  .name(configuration.name)
-  .version(configuration.version)
-  .description(configuration.description)
+  .name(config.name)
+  .version(config.version)
+  .description(config.description)
   .action(() => console.log("Hello World, my name is Contribution Mate!"))
-  .command("config", configCommand)
+  .command("config", configCmd)
   .parse(Deno.args);
